Development

+++ API_NEST/API_ADA_COMPANY_NESTJS/src/application/use-cases/cliente/delete-cliente.use-case.ts
@@ -1,9 +1,13 @@
-import { ClienteRepositoryImpl } from '../../../infrastructure/database/repositories/cliente.repository';
+import { HttpException, HttpStatus } from '@nestjs/common';
 
 export class DeleteClienteUseCase {
-  constructor(private readonly clienteRepository: ClienteRepositoryImpl) {}
+  constructor(private readonly clienteRepository: any) {}
 
   async execute(id: string): Promise<void> {
+    const cliente = await this.clienteRepository.findById(id);
+    if (!cliente) {
+      throw new HttpException('Cliente não encontrado', HttpStatus.NOT_FOUND);
+    }
     await this.clienteRepository.delete(id);
   }
 }
